Arka/requirements.txt
discord.py==2.3.2
pymongo==4.5.0
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.0
python-dotenv==1.0.0
numpy==1.26.2
pydantic==2.5.2